fix: testing values clean up
